This commit was generated by cvs2svn to compensate for changes in r11913, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@11914 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/vicii/vicii-badline.c
@@ -166,9 +166,9 @@ void vicii_badline_check_state(BYTE value, int cycle, unsigned int line,
 
     /* Check whether bad line state has changed.  */
     was_bad_line = (old_allow_bad_lines
-                    && (vic_ii.raster.ysmooth == (line & 7)));
+                    && (vic_ii.raster.ysmooth == (int)(line & 7)));
     now_bad_line = (vic_ii.allow_bad_lines
-                    && ((value & 7) == (line & 7)));
+                    && ((int)(value & 7) == (int)(line & 7)));
 
     if (was_bad_line && !now_bad_line) {
         line_becomes_good(cycle);
